move dto to shared lib
