+++ project/apps/blog/src/app/publication/publication.constant.ts
@@ -5,10 +5,7 @@ export const  API_TAG_NAME ='publications'
 export const PublicationsError = {
   PublicationNotFound : 'Publication is not found',
   Delete : 'Publication is not deleted',
-  WrongSourse : 'Video link must be from Youtube',
   WrongType : 'Wrong post type',
-  WrongTagStart : 'Tags should start with digit',
-  SpacesInTag:'Tags should not contain spaces',
   AlreadyReposted:'You already reposted this publication',
   NotUserAuthor:'User is not an author of this publication',
 } as const;
@@ -27,37 +24,11 @@ export const PublicationPath = {
   Id: ':id'
 }as const;
 
-export const PublicationTitleLength = {
-  Min: 20,
-  Max: 50,
-};
-export const PublicationAnnouncementLength = {
-  Min: 50,
-  Max: 225,
-};
-export const TextPublicationLength = {
-  Min: 100,
-  Max: 1024,
-};
-export const QuotePublicationLength = {
-  Min: 20,
-  Max: 300,
-};
-export const PublicationAuthorLength = {
-  Min: 3,
-  Max: 50,
-};
+
+export const VALIDATION_ARGUMENT_TYPE = 'body'
+
 export const TagDefaultParam = {
   MinLength: 3,
   MaxLength: 10,
   Amount: 8,
 };
-
-export const LINK_DESCRIPTION_LENGTH = 300;
-
-export const RegExpPattern = {
-  Video:/(?:(?:https?:\/\/)(?:www)?\.?(?:youtu\.?be)(?:\.com)?\/(?:.*[=/])*)([^= &?/\r\n]{8,11})/,
-  Tag:/^[A-Za-zА-Яа-я]([A-Za-zА-Яа-я0-9_.])+$/g,
-} ;
-
-export const VALIDATION_ARGUMENT_TYPE = 'body'
