+++ project/apps/blog/src/app/publication/publication.controller.ts
@@ -3,8 +3,7 @@ import { PublicationService } from './publication.service';
 import { ApiResponse, ApiTags } from '@nestjs/swagger';
 import { API_TAG_NAME, PublicationMessages, PublicationPath, PublicationsError } from './publication.constant';
 import { PublicationRdo } from './rdo/publication.rdo';
-import { CreateBlogPublicationDto } from './dto/create/blog-publication-dto.type';
-import { UpdateBlogPublicationDto } from './dto/update/blog-publication-dto.type';
+import { CreateBlogPublicationDto,UpdateBlogPublicationDto } from '@project/shared/shared-dto';
 import { adaptRdoPublication } from './utils/adapt-rdo-publication';
 import { CreatePostValidationPipe } from './pipes/create-post-validation.pipe';
 import { UpdatePostValidationPipe } from './pipes/update-post-validation.pipe';
