--- project/apps/users/src/app/authentication/dto/login-user.dto.ts
@@ -1,20 +1,0 @@
-import { ApiProperty } from '@nestjs/swagger';
-import { IsEmail, IsString, MaxLength, MinLength } from 'class-validator';
-import { AuthError, UserPasswordLength } from '../authentication.constant';
-
-export class LoginUserDto {
-  @ApiProperty({
-    description: 'User email address',
-    example: 'test@test.ru'
-  })
-  @IsEmail({}, { message: AuthError.InvalidEmail})
-  public email: string;
-  @ApiProperty({
-    description: 'User password',
-    example: '123456'
-  })
-  @IsString()
-  @MinLength(UserPasswordLength.Min)
-  @MaxLength(UserPasswordLength.Max)
-  public password: string;
-}
