+++ project/libs/shared/shared-dto/README.md
@@ -1,0 +1,7 @@
+# shared-shared-dto
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test shared-shared-dto` to execute the unit tests via [Jest](https://jestjs.io).
