+++ project/libs/shared/shared-dto/src/index.ts
@@ -1,0 +1,18 @@
+export * from './lib/user/change-password.dto';
+export * from './lib/user/create-user.dto';
+export * from './lib/user/login-user.dto';
+export * from './lib/comment/create-comment.dto';
+export * from './lib/publications/create/blog-publication-dto.type';
+export * from './lib/publications/create/publication-link.dto';
+export * from './lib/publications/create/publication-photo.dto';
+export * from './lib/publications/create/publication-quote.dto';
+export * from './lib/publications/create/publication-text.dto';
+export * from './lib/publications/create/publication-video.dto';
+export * from './lib/publications/create/publication.dto';
+export * from './lib/publications/update/blog-publication-dto.type';
+export * from './lib/publications/update/publication-link.dto';
+export * from './lib/publications/update/publication-photo.dto';
+export * from './lib/publications/update/publication-quote.dto';
+export * from './lib/publications/update/publication-text.dto';
+export * from './lib/publications/update/publication-video.dto';
+export * from './lib/publications/update/publication.dto';
