+++ project/libs/shared/shared-dto/src/lib/comment/comment-dto.constant.ts
@@ -1,0 +1,4 @@
+export const CommentTextLength = {
+  Min: 10,
+  Max: 300,
+};
