+++ project/libs/shared/shared-dto/src/lib/publications/create/blog-publication-dto.type.ts
