+++ project/libs/shared/shared-dto/src/lib/publications/create/publication-photo.dto.ts
