+++ project/libs/shared/shared-dto/src/lib/publications/create/publication-quote.dto.ts
@@ -1,26 +1,25 @@
 import { ApiProperty } from "@nestjs/swagger";
-import { UpdatePublicationDto } from "./publication.dto";
-import { IsOptional, IsString, MaxLength, MinLength } from "class-validator";
-import { PublicationAuthorLength, QuotePublicationLength } from "../../publication.constant";
+import { CreatePublicationDto } from "./publication.dto";
+import { IsString, MaxLength, MinLength } from "class-validator";
+import { PublicationAuthorLength, QuotePublicationLength } from "../publication-dto.constant";
 
-export class UpdateQuotePublicationDto extends UpdatePublicationDto {
+export class CreateQuotePublicationDto extends CreatePublicationDto {
   @ApiProperty({
     description: 'Text of publication  quote',
     example: 'Quote'
   })
   @IsString()
-  @IsOptional()
   @MinLength(QuotePublicationLength.Min)
   @MaxLength(QuotePublicationLength.Max)
-  public text?: string;
+  public text: string;
 
   @ApiProperty({
     description: 'Author name',
     example: 'Author'
   })
   @IsString()
-  @IsOptional()
   @MinLength(PublicationAuthorLength.Min)
   @MaxLength(PublicationAuthorLength.Max)
-  public author?: string;
+  public author: string;
+
 }
