+++ project/libs/shared/shared-dto/src/lib/publications/update/blog-publication-dto.type.ts
