+++ project/libs/shared/shared-dto/src/lib/publications/update/publication-photo.dto.ts
