+++ project/libs/shared/shared-dto/src/lib/publications/update/publication-text.dto.ts
@@ -1,31 +1,35 @@
 import { ApiProperty } from "@nestjs/swagger";
-import { CreatePublicationDto } from "./publication.dto";
-import { PublicationAnnouncementLength, PublicationTitleLength, TextPublicationLength } from "../../publication.constant";
-import { IsString, MaxLength, MinLength } from "class-validator";
-
-export class CreateTextPublicationDto extends CreatePublicationDto{
+import { UpdatePublicationDto } from "./publication.dto";
+import { PublicationAnnouncementLength, PublicationTitleLength, TextPublicationLength } from "../publication-dto.constant";
+import { IsOptional, IsString, MaxLength, MinLength } from "class-validator";
+export class UpdateTextPublicationDto extends UpdatePublicationDto{
   @ApiProperty({
     description: 'Text of post',
     example: 'Text'
   })
   @IsString()
+  @IsOptional()
   @MinLength(TextPublicationLength.Min)
   @MaxLength(TextPublicationLength.Max)
-  public text: string;
+  public text?: string;
+
   @ApiProperty({
     description: 'Title of post',
     example: 'Title'
   })
   @IsString()
+  @IsOptional()
   @MinLength(PublicationTitleLength.Min)
   @MaxLength(PublicationTitleLength.Max)
-  public title: string;
+  public title?: string;
+
   @ApiProperty({
     description: 'Announcement of post',
     example: 'Text'
   })
   @IsString()
+  @IsOptional()
   @MinLength(PublicationAnnouncementLength.Min)
   @MaxLength(PublicationAnnouncementLength.Max)
-  public announcement: string;
+  public announcement?: string;
 }
