+++ project/libs/shared/shared-dto/src/lib/publications/update/publication-video.dto.ts
@@ -1,7 +1,7 @@
 import { ApiProperty } from "@nestjs/swagger";
 import { UpdatePublicationDto } from "./publication.dto";
 import { IsOptional, IsString, IsUrl, Matches, MaxLength, MinLength } from "class-validator";
-import { PublicationTitleLength, PublicationsError, RegExpPattern } from "../../publication.constant";
+import { PublicationTitleLength, PublicationsError, RegExpPattern } from "../publication-dto.constant";
 
 export class UpdateVideoPublicationDto extends UpdatePublicationDto{
   @ApiProperty({
