+++ project/libs/shared/shared-dto/src/lib/user/change-password.dto.ts
@@ -1,6 +1,6 @@
 import { ApiProperty } from "@nestjs/swagger";
 import { IsString, MaxLength, MinLength } from "class-validator";
-import { UserPasswordLength } from "../app.constant";
+import { UserPasswordLength} from './user-dto.constant'
 
 export class ChangePasswordDto {
