+++ project/libs/shared/shared-dto/src/lib/user/user-dto.constant.ts
@@ -1,0 +1,11 @@
+export const EMAIL_ERROR = 'Email is invalid';
+
+export const UserNameLength = {
+  Min: 3,
+  Max: 50,
+};
+
+export const UserPasswordLength = {
+  Min: 6,
+  Max: 12,
+};
