+++ project/tsconfig.base.json
@@ -25,6 +25,7 @@
       "@project/config/config-users": ["libs/config/config-users/src/index.ts"],
       "@project/models/blog-models": ["libs/models/blog-models/src/index.ts"],
       "@project/shared/app-types": ["libs/shared/app-types/src/index.ts"],
+      "@project/shared/shared-dto": ["libs/shared/shared-dto/src/index.ts"],
       "@project/shared/shared-pipes": ["libs/shared/shared-pipes/src/index.ts"],
       "@project/util/util-core": ["libs/util/util-core/src/index.ts"],
       "@project/util/util-types": ["libs/util/util-types/src/index.ts"]
